pyproject.toml
[project]
name = "perceptual-straightening-in-video-diffusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diffusers>=0.37.1",
    "huggingface>=0.0.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "tqdm>=4.67.3",
    "transformers>=5.5.3",
    "xformers>=0.0.35",
]
